[REFACTOR] make password input an abstract component
