+++ src/components/atoms/__tests__/input.test.tsx
@@ -37,4 +37,26 @@ describe('<Input />', () => {
     expect(ref.current).toBeInTheDocument()
     expect(ref.current).toBeInstanceOf(HTMLInputElement)
   })
+
+  it('should toggle password visibility when the eye icon is clicked', () => {
+    render(<Input type="password" placeholder="Password" />)
+    const passwordInput = screen.getByPlaceholderText('Password')
+    const toggleButton = screen
+      .getAllByRole('button')
+      .find(button => button.querySelector('svg'))
+    expect(toggleButton).toBeInTheDocument()
+
+    // Initially, the password should be hidden
+    expect(passwordInput).toHaveAttribute('type', 'password')
+
+    // Click to show the password
+    if (toggleButton) {
+      fireEvent.click(toggleButton)
+      expect(passwordInput).toHaveAttribute('type', 'text')
+
+      // Click again to hide the password
+      fireEvent.click(toggleButton)
+      expect(passwordInput).toHaveAttribute('type', 'password')
+    }
+  })
 })
